Move linebuilder into its own class
